+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.0] - 2025-08-26
+### Added
+- Update Data table properties method
+- Data table rows bulk replace method
+
 ## [0.13.0] - 2025-08-18
 ### Added
 - Find UDM Field Values functionality